ASB sends a command MSMQ and subscribes to an event published by MSMQ
