+++ samples/azure/azure-service-bus-msmq-bridge/ASB_7/Bridge/Program.cs
@@ -8,6 +8,8 @@ class Program
 {
     static async Task Main(string[] args)
     {
+        Console.Title = "Samples.Azure.ServiceBus.Bridge";
+
         var connectionString = Environment.GetEnvironmentVariable("AzureServiceBus.ConnectionString");
         if (string.IsNullOrWhiteSpace(connectionString))
         {
