+++ samples/azure/azure-service-bus-msmq-bridge/ASB_7/MsmqEndpoint/MsmqEndpoint.csproj
@@ -8,7 +8,6 @@
 
   <ItemGroup>
     <PackageReference Include="NServiceBus" Version="6.*" />
-    <PackageReference Include="NServiceBus.Bridge.Connector" Version="1.*" />
   </ItemGroup>
   
   <ItemGroup>
