+++ samples/azure/azure-service-bus-msmq-bridge/ASB_7/MsmqEndpoint/Program.cs
@@ -13,51 +13,16 @@ class Program
         endpointConfiguration.EnableInstallers();
         endpointConfiguration.UsePersistence<InMemoryPersistence>();
         endpointConfiguration.UseSerialization<XmlSerializer>();
+        endpointConfiguration.AddDeserializer<JsonSerializer>();
         var recoverability = endpointConfiguration.Recoverability();
         recoverability.DisableLegacyRetriesSatellite();
-        var transport = endpointConfiguration.UseTransport<MsmqTransport>();
-
-        #region connect-msmq-side-of-bridge
-
-        var routing = transport.Routing();
-        var bridge = routing.ConnectToBridge("Bridge-MSMQ");
-
-        #endregion
-
-        #region route-command-via-bridge
-
-        bridge.RouteToEndpoint(typeof(MyCommand), "Samples.Azure.ServiceBus.AsbEndpoint");
-
-        #endregion
-        
+        endpointConfiguration.UseTransport<MsmqTransport>();
+       
         var endpointInstance = await Endpoint.Start(endpointConfiguration)
             .ConfigureAwait(false);
 
-        Console.WriteLine("Press 'c' to send a command");
-        Console.WriteLine("Press 'e' to publish an event");
-        Console.WriteLine("Press any other key to exit");
-        var @continue = true;
-
-        while (@continue)
-        {
-            var key = Console.ReadKey().Key;
-            switch (key)
-            {
-                case ConsoleKey.C:
-                    await endpointInstance.Send(new MyCommand { Property = "command from MSMQ endpoint"}).ConfigureAwait(false);
-                    Console.WriteLine("\nCommand sent");
-                    break;
-
-                case ConsoleKey.E:
-                    await endpointInstance.Publish<MyEvent>(@event => { @event.Property = "event from MSMQ endpoint"; }).ConfigureAwait(false);
-                    Console.WriteLine("\nEvent sent");
-                    break;
-
-                default:
-                    @continue = false;
-                    break;
-            }
-        }
+        Console.WriteLine("Press any key to exit");
+        Console.ReadKey();
 
         await endpointInstance.Stop()
             .ConfigureAwait(false);
